requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
numpy==1.26.4
supabase==2.11.0
python-dotenv==1.0.1
httpx==0.27.0
streamlit==1.32.0
plotly==5.19.0
pydantic==2.6.3
python-multipart==0.0.9
openai==1.13.3
